# Unroll the zeroing loop