Zeroing the stack is a huge part of the total cost of cross-compartment calls.  Running on Sail (reporting retired instructions), the compartment-switcher benchmark (from #37) reports (stack size, call+return, call, return):

0x100   213     127     86
0x200   405     223     182
0x400   789     415     374
0x800   1557    799     758
0x1000  3093    1567    1526

If we skip stack zeroing, the numbers are 95, 67, 28 for all of them. This means that, even with tiny stacks, zeroing is more than half of the cost of a compartment switch and that cost grows with larger compartments.

We currently require three instructions per store in the zeroing loop. This commit unrolls the loop so that we zero 32 bytes at a time.  The results are now:

0x100   185     113     72
0x200   313     177     136
0x400   569     305     264
0x800   1081    561     520
0x1000  2305    1073    1032

512 bytes is probably the smallest stack size that makes sense for a thread and here we see almost a 25% speedup. This has a fairly noticeable impact on the test suite too:

Before:
Test runner: All tests finished in 2385962 cycles
After:
Test runner: All tests finished in 2087477 cycles